Revert "Horizonal forecast layout"
This reverts commit a8ebaa0ecfa5de055771a2f429080727609c9389.

+++ app/src/main/java/com/revengeos/weather/response/Hourly.java
@@ -30,5 +30,5 @@ public class Hourly {
     @SerializedName("weather")
     public List<Weather> weather = null;
     @SerializedName("pop")
-    public float pop;
+    public int pop;
 }
